+++ sdk/cpp/core/src/restconf_client.hpp
@@ -12,6 +12,11 @@
  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  See the License for the specific language governing permissions and
  limitations under the License.
+ -------------------------------------------------------------------
+ This file has been modified by Yan Gorelik, YDK Solutions.
+ All modifications in original under CiscoDevNet domain
+ introduced since October 2019 are copyrighted.
+ All rights reserved under Apache License, Version 2.0.
 ------------------------------------------------------------------*/
 
 #ifndef _RESTCONF_CLIENT_H_
@@ -19,17 +24,21 @@
 
 #include <string>
 
-
 typedef void CURL;
 struct curl_slist;
 
 namespace ydk
 {
+
+// Flags to control SSL verification
+#define SSL_PEER_VERIFICATION 0x01
+#define SSL_HOSTNAME_VERIFICATION 0x02
+
 class RestconfClient
 {
     public:
         RestconfClient(const std::string & address, const std::string & username, const std::string & password,
-                    int port, const std::string & encoding);
+                    int port, const std::string & encoding, uint ssl_options=0);
         ~RestconfClient();
 
         std::string execute(const std::string & yfilter, const std::string & url, const std::string & payload) const;
@@ -45,6 +54,8 @@ class RestconfClient
         curl_slist *header_options_list;
         std::string base_url;
         std::string encoding;
+        std::string protocol;
+        uint ssl_options;
 };
 }