+++ sdk/go/core/ydk/types/types.go
@@ -1,22 +1,25 @@
-// Package types provides built-in types specified in
-// YANG RFC 6020 and types used in YDK Go APIs.
-//
-// YANG Development Kit Copyright 2017 Cisco Systems. All rights reserved.
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+/*  ----------------------------------------------------------------
+ YDK - YANG Development Kit
+ Copyright 2016 Cisco Systems. All rights reserved.
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+ http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+ -------------------------------------------------------------------
+ This file has been modified by Yan Gorelik, YDK Solutions.
+ All modifications in original under CiscoDevNet domain
+ introduced since October 2019 are copyrighted.
+ All rights reserved under Apache License, Version 2.0.
+ ------------------------------------------------------------------*/
+
 package types
 
 import (
@@ -536,7 +539,10 @@ func getLeafValue(value interface{}) LeafData {
 	var leafData LeafData
 	switch value.(type) {
 	case yfilter.YFilter:
-		leafData = LeafData{IsSet: true, Filter: value.(yfilter.YFilter)}
+		leafData = LeafData{IsSet: false, Filter: value.(yfilter.YFilter)}
+	case LeafData:
+	    leafData = value.(LeafData)
+	    leafData.IsSet = true
 	case map[string]bool:
 		// bits
 		var used_bits []string
@@ -592,7 +598,6 @@ func GetEntityPath(entity Entity) EntityPath {
 				entityPath.ValuePaths = append(
 					entityPath.ValuePaths,
 					NameLeafData{Name: path, Data: leafData})
-				if leafData.Filter != yfilter.NotSet { break }
 			}
 		}
 	}
@@ -833,8 +838,8 @@ func GetRelativeEntityPath(current_node Entity, ancestor Entity, path string) st
 
 }
 
-// IsSet returns whether the given filter is set or not
-func IsSet(Filter yfilter.YFilter) bool {
+// IsFilterSet returns whether the given filter is set or not
+func IsFilterSet(Filter yfilter.YFilter) bool {
 	return Filter != yfilter.NotSet
 }