+++ ydkgen/__init__.py
@@ -13,10 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ------------------------------------------------------------------
+# This file has been modified by Yan Gorelik, YDK Solutions.
+# All modifications in original under CiscoDevNet domain
+# introduced since October 2019 are copyrighted.
+# All rights reserved under Apache License, Version 2.0.
+# ------------------------------------------------------------------
+
 """
 Generate YDK core library, bundle packages.
 """
-from __future__ import print_function
+
 from distutils import dir_util
 
 import fileinput
@@ -148,7 +154,7 @@ class YdkGenerator(object):
         generated_files = self._print_packages(packages, gen_api_root, curr_bundle)
 
         if self.language == 'cpp':
-            _modify_cpp_cmake(gen_api_root, curr_bundle.name, cpp_version, cpp_build, 
+            _modify_cpp_cmake(gen_api_root, curr_bundle.name, cpp_version, cpp_build,
                               curr_bundle.str_core_version, generated_files[0], generated_files[1], yang_models)
 
         os.remove(tmp_file)